Add helper mock fetch method for DateTimeSelectPage tests
